[DSPDC-1763] Autogenerate HCA schema when creating dataset (#133)
Why

We need to auto-generate the HCA schema when creating an HCA dataset, rather than relying on pulling it down via a well-known gist every time.

This PR

My original plan was to just glob up the schema JSON files from the schema module in HCA. However, those schemas are not in a format consumable by Jade. As part of our build process, we apply a scala plugin that does the needed transformation. I've decided to delegate the schema generation to that task and run it via a subprocess.run() call if the needed schema json is not present. This isn't going to be great for CI, but it establishes a generate_schema interface that we can iterate on and allows us to proceed with the creation of the needed E2E test which this ticket is positioned to enable.
Refactors the argument handling to use subparsers rather than argument groups. Argument groups are awkward for the purposes of this script (i.e., both "create" and "query" consume the dataset name argument, but we were re-purposing the remove version which is confusing. A subparser gives us a cleaner separate between the 3 modes of this script.
Extracts the _exec_tdr_operation wrapper to a formal decorator in .common which we can reuse if need be.