[DSPDC-1876] Fix CI slowness by setting request timeout on fixture dataset deletion (#173)
Why

CI is taking upwards of 30 minutes to complete. I've tracked this down to the fixture dataset deletion request we are making to TDR. Specifically, there is no timeout set in our calls to the client, so I believe urllib3 under the hood is using the default system socket timeouts. It's still unclear why this call would take minutes.

This PR

Adds a 30 second timeout to the delete_dataset call. This causes a quick failure and then a retry which immediately succeeds.