[DSPDC-1879] Tool to find project row IDs in TDR (#177)
Why

There was an issue in DCP8 where a project was included for import that should not have been. The fix in these cases is to
find the related datarepo_row_ids in the project and links tables and submit them for soft deletion. This is an attempt
to script up the first part of that process so we don't have to adjourn to the BQ web console every time this happens.

This PR

Adds a find_project_rows.py script that outputs the project and links datarepo_row_ids in question.