[DSPDC-1890] Add partitioning logic (#184)
Why

Relevant ticket
We should use dagster's partitioning backfill feature to let it automate the load of the 137 projects for the dev refresh.

This PR

Puts the list of projects in a CSV that we parse and hand to dagster as our definition of the partition set
Adds a mostly hardcoded run config which we will need to adjust as dataset's change etc. I'm not particularly fond of the hardcoding of this config as it'll require a PR + deploy if any part of the config changes (i.e, the target dataset for instance).