[DSPDC-1814] Per project dataset copying (#183)
Why

We would like a dataset + snapshot per project when copying.

This PR

Adds a slightly different flavor of the copy_project pipeline, copy_project_to_new_dataset. This pipeline checks for the existence of a dataset for the supplied project_id, and if not present creates it. Data is then copied to this new dataset.
Migrates the copy_project code to the new dagster concepts of graph, ops and jobs (roughly analogous to pipeline, solid and mode).