[DSPDC-1898] Add partitioning support DCP releases (#206)
Why

We would like to leverage Dagster's backfill feature to better automate DCP imports.

This PR

Adds code that looks for a manifest file in google storage at the DCP_RELEASE_MANIFEST_PATH env var. It scans the path for files of the form <prefix>_manifest.csv. Each file is assumed to contain a CSV of GS staging paths. The release process with this mechanism will look like:
The importing engineer/goalie will grab the manifest CSV from the google sheet and place at gs://broad-dsp-monster-hca-dev-dcp-releases/manifests/dcpXYZ_manifest.csv.
Reload the repo in prod
Run a backfill for the newly discovered partitions from the manifest that was just placed in the bucket
Updates the helmfile config to put the DCP_RELEASE_MANIFEST_PATH in the environment