Un-hardcode BQ location (#221)
Why

We have datasets in dev that are not always in the US multi-region. We should be able to import to dataset's in any region needed.

This PR

Un-hardcodes the US multi-region
Determines the region of the target dataset at runtime and uses that to a) determine where to create the "staging" dataset and b) thread through the location to downstream import queries
The staging bucket must also be located in the same region. For now, I've created a staging bucket in "us-central1" that we can use. We'll need to determine if there's a better approach than a staging bucket per region.