Add release date cutoff param to manifest verfication tool (#229)
Why

We verify that the data present in a staging area has been loaded properly by running the verify_release_manifest.py tool. However, if an updated file is staged after an ingest, the verification tool will fail because the content will not match. We should guard against this situation.

This PR

Adds a release_cutoff parameter to the tool that will ignore any files staged after the provided timestamp