Subgraph verification tool (#234)
Why

There was an issue in dev where invalid subgraphs were loaded and handed off downstream. We should be able to verify the consistency of the subgraphs loaded in a dataset (i.e, that all nodes referred to by the subgraph are present in the dataset).

This PR

Adds a tool that will inspect the subgraphs in a dataset and verify that all of the entity IDs are loaded in the target dataset
Refactors the node hydration code to a subgraphs.py module that is now shared between the project copying pipeline and this tool.