[DSPDC-1930] E2E testing for copy project pipeline (#258)
Why

We need an end-to-end test in place for the copy project pipeline before we can proceed with the formal "new" production migration.

This PR

Adds the needed test. It loads a new HCA dataset in dev, snapshots it, copies the constituent project to a new dataset and asserts that the data was copied properly.