[DSPDC-1949] Prod migration partitioning and validation (#263)
Why

We are going to be copying ~250 projects from HCA prod to new prod. To ease this transition, we'll want to leverage partitioning mechanism to help us track which project's have been migrated and which have not.

This PR

Adds new dcp1/dcp2_copy_project jobs and associated run configs. I was hoping that the copy_project_from_x_to_y would be flexible enough to be the single job for this process, but dcp1 resides in a different snapshot. I judged the complexity involved in supporting multiple source datasets not worth the effort, hence the one-offs.
Added slack_notification sensors, similar to what we do in DAP
Added a verify_subgraph solid that we invoke after the validate_dataset solid. The idea here is to hydrate the copied subgraphs and ensure all descendant nodes are present.