[DSPDC-1949] Direct data file ingestion (#264)
Why

We have very large projects in HCA prod that cause GCS rate limiting. Jade has set the permission on the source bucket for the
offending projects so we can perform a direct bulk file ingest out of TDR, rather than having to copy to a staging area first.

This PR

Adds a config schema value that will switch on direct copy by default. We still need to be able to turn it off in most cases as this permission will not be set everywhere